Refactor code to have all core building blocks in one file
