+++ src/extract-left.mjs
@@ -1,4 +1,4 @@
-import {application1, future} from './internal/check';
+import {application1, future} from './future';
 
 export function extractLeft(m){
   application1(extractLeft, future, arguments);
