+++ src/extract-right.mjs
@@ -1,4 +1,4 @@
-import {application1, future} from './internal/check';
+import {application1, future} from './future';
 
 export function extractRight(m){
   application1(extractRight, future, arguments);
