--- src/internal/transformation.mjs
@@ -1,72 +1,0 @@
-import {invalidFuture} from './error';
-import {nil} from './list';
-import {noop, moop, show} from './utils';
-
-import {Crash} from '../crash';
-import {isFuture} from '../future';
-import {Reject} from '../reject';
-import {Resolve} from '../resolve';
-
-function BaseTransformation$rejected(x){
-  this.cancel();
-  return new Reject(this.context, x);
-}
-
-function BaseTransformation$resolved(x){
-  this.cancel();
-  return new Resolve(this.context, x);
-}
-
-export var BaseTransformation = {
-  rejected: BaseTransformation$rejected,
-  resolved: BaseTransformation$resolved,
-  run: moop,
-  cancel: noop,
-  context: nil,
-  arity: 0,
-  name: 'transform'
-};
-
-function wrapHandler(handler){
-  return function transformationHandler(x){
-    var m;
-    try{
-      m = handler.call(this, x);
-    }catch(e){
-      return new Crash(this.context, e);
-    }
-    if(isFuture(m)){
-      return m;
-    }
-    return new Crash(this.context, invalidFuture(
-      this.name + ' expects the return value from the function it\'s given', m,
-      '\n  When called with: ' + show(x)
-    ));
-  };
-}
-
-export function createTransformation(arity, name, prototype){
-  var Transformation = function(context, $1, $2){
-    this.context = context;
-    this.$1 = $1;
-    this.$2 = $2;
-  };
-
-  Transformation.prototype = Object.create(BaseTransformation);
-  Transformation.prototype.arity = arity;
-  Transformation.prototype.name = name;
-
-  if(typeof prototype.rejected === 'function'){
-    Transformation.prototype.rejected = wrapHandler(prototype.rejected);
-  }
-
-  if(typeof prototype.resolved === 'function'){
-    Transformation.prototype.resolved = wrapHandler(prototype.resolved);
-  }
-
-  if(typeof prototype.run === 'function'){
-    Transformation.prototype.run = prototype.run;
-  }
-
-  return Transformation;
-}
