+++ src/value.mjs
@@ -1,6 +1,6 @@
-import {application1, application, func, future} from './internal/check';
 import {error} from './internal/error';
 import {raise, show} from './internal/utils';
+import {application1, application, func, future} from './future';
 
 export function value(res){
   var context1 = application1(value, func, arguments);
