+++ test/unit/5.chain-rec.mjs
@@ -1,6 +1,6 @@
 import chai from 'chai';
 import {resolve, after, reject} from '../../index.mjs';
-import {chainRec} from '../../src/chain-rec';
+import {chainRec} from '../../src/future';
 import {isIteration} from '../../src/internal/iteration';
 import {assertCrashed, assertRejected, assertResolved, error, failRej, failRes, noop} from '../util/util';
